When looking up taflmen, verify we aren't asking for a dead taflman
In TaflmanCoordMap:
1. When a taflman is removed, we set its mTaflmen entry to 0. This is
Taflman.EMPTY, but it also matches a taflman with ID 0 (that is, usually the
king).
2. This usually doesn't matter. The piece that gets replaced with Taflman.EMPTY
always comes after the piece with ID 0, because the piece with ID 0 is first
in the array. When we ask to get a piece with ID 0, we return the first one.
3. It does matter when we use getTaflman, as in the board array generation code
in BoardImpl, because the dummy piece will end up coming second, and probably
overwriting the actual piece with ID 0.
4. Check to be sure that not only the IDs match, but that the whole encoded
taflman matches.