UI overhaul (in one commit, I know)
The UI is now organized much more neatly. From the user view down:

1. Windows
Windows are UI elements, and belong to screens.

2. Screens
Screens are containers and managers for UI elements, and contain windows. They
have a TerminalCallback.

Screens may be one user-viewable component, such as the in-game screen, or
multiple user-viewable components which go in and out of visibility, such as the
menu screen.

3. Terminal
AdvancedTerminal is the OpenTafl UI-level framework.

--HG--
branch : v0.2.x