Use entering move as identifying move for game states
This is good, because the entering move is the unique identifier for the tree.
This is bad, because the code assumes in a lot of places that the exiting move
is the move being used.
This is okay, because I think I've solved it. (Mostly.)
Still to fix:
1. Commands which take a move address
2. Saved games spitting out an extra thingy

--HG--
branch : puzzles