Fixes for some extension search bugs
1. Time use: be willing to jump out of continuation search if the clock is
running down, now that we have cached values to fall back on if we haven't
finished a search.

2. Cache values during extension searches: if the extension search fails and we
have a good cached value, restore it. (If we don't have a good cached value,
that means we're past the known part of the tree, and it's useful information
to evaluate the current state and return that.)

3. Set a meaningful overallMaxDepth for horizon search: this allows us to do
the above decision.

4. Correctly reset the accumulator variables in FishyEvaluator: this caused
states evaluated later to evaluate strangely, because some of the variables
would already be pegged.

5. Consult the repetition table during continuation searches: required so we
know to avoid continuation searches on losing repetitions.

--HG--
branch : ai-improvements